+++ changelog.md
@@ -11,6 +11,7 @@
 ### Improvements
 - added time field to game form
 - lfg role management now uses modals for better user experience
+- /draft can now seat players when used within an lfg thread
 
 ### Bug Fixes
 - fixed bot cold start issues with slow response time